tests(integration): separate snapshot tests into own module
